Module clamor::chain_spec

A chain specification, or “chain spec”, is a collection of configuration information that dictates which network a blockchain node will connect to, which entities it will initially communicate with, and what consensus-critical state it must have at genesis.

Source: https://docs.substrate.io/v3/runtime/chain-specs/

Functions

authority_keys_from_seed

Generate an Aura authority key.

development_config

get_account_id_from_seed

Generate an account ID from seed.

get_from_seed

Generate a crypto pair from seed.

get_from_seed_to_eth

Generate a crypto pair from seed.

live_config

local_testnet_config

Type Definitions

ChainSpec

Specialized ChainSpec. This is a specialization of the general Substrate ChainSpec type.

DetachId

EthId

UploadId